requirements.txt
ddddocr~=1.5.6
requests~=2.32.3
uvicorn~=0.34.3
pycryptodome~=3.23.0
fake-useragent~=2.2.0
fastapi~=0.115.11
pydantic~=2.10.6